requirements.txt
alembic==1.14.0
blinker==1.9.0
click==8.1.7
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.0
Flask-DebugToolbar==0.16.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
mysql-connector-python==8.1.0
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
plotly==5.24.1
protobuf==4.21.12
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.36
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
